requirements.txt
azure-functions
azure-core
azure-search-documents==11.6.0
openai>=1.35.0
azure-identity